ファイルを分割する
